Move from retrieve to chatbot
