+++ ai-services/assets/applications/rag-dev/openshift/templates/backend-deployment.yaml
@@ -28,7 +28,7 @@ spec:
             - "/var/venv/bin/python"
             - "-m"
             - "uvicorn"
-            - "retrieve.app:app"
+            - "chatbot.app:app"
             - "--host"
             - "0.0.0.0"
             - "--port"
