+++ ai-services/assets/applications/rag-dev/openshift/values.yaml
@@ -4,19 +4,19 @@ ui:
 
 backend:
   # @hidden
-  image: icr.io/ai-services-cicd/rag:v0.0.58
+  image: icr.io/ai-services-cicd/rag:v0.0.59
   # @hidden
   log_level: "INFO"
 
 summarize:
   # @hidden
-  image: icr.io/ai-services-cicd/rag:v0.0.58
+  image: icr.io/ai-services-cicd/rag:v0.0.59
   # @hidden
   log_level: "INFO"
 
 digitize:
   # @hidden
-  image: icr.io/ai-services-cicd/rag:v0.0.58
+  image: icr.io/ai-services-cicd/rag:v0.0.59
   # @hidden
   log_level: "INFO"
