+++ ai-services/assets/applications/rag-dev/podman/values.yaml
@@ -8,7 +8,7 @@ backend:
   # @description Host port for the OpenAI-compatible RAG service. Defaults to unexposed; assign a port to enable external access.
   port: "0"
   # @hidden
-  image: icr.io/ai-services-cicd/rag:v0.0.58
+  image: icr.io/ai-services-cicd/rag:v0.0.59
   # @hidden
   log_level: "INFO"
 
@@ -16,7 +16,7 @@ digitize:
   # @description Host port for the DIGITIZE API. If unspecified, a random available port is assigned. Specify a port number to use a custom value.
   port: ""
   # @hidden
-  image: icr.io/ai-services-cicd/rag:v0.0.58
+  image: icr.io/ai-services-cicd/rag:v0.0.59
   # @hidden
   log_level: "INFO"
 
@@ -49,6 +49,6 @@ summarize:
   # @description Host port for the Summarize API. If unspecified, a random available port is assigned. Specify a port number to use a custom value.
   port: "6000"
   # @hidden
-  image: icr.io/ai-services-cicd/rag:v0.0.58
+  image: icr.io/ai-services-cicd/rag:v0.0.59
   # @hidden
   log_level: "INFO"
