+++ ai-services/assets/applications/rag/podman/templates/chat-bot.yaml.tmpl
@@ -39,7 +39,7 @@ spec:
         - "/var/venv/bin/python"
         - "-m"
         - "uvicorn"
-        - "retrieve.app:app"
+        - "chatbot.app:app"
         - "--host"
         - "0.0.0.0"
         - "--port"
