+++ spyre-rag/src/chatbot/readme.md
@@ -2,4 +2,4 @@
 
 pip install -r requirements.txt --extra-index-url=https://wheels.developerfirst.ibm.com/ppc64le/linux
 
-python -m retrieve.app
+python -m chatbot.app
